Move de- & serializeSettings into its own module
